Massive high energy commit
When moving the codebase I fucked up. Here's all changes in one big af package:
* Cleanup
* Short code review
* Fix wrong Q value calculation
* Optimize P (remove redundant nested hashmap)
* Make all floats double
* Add DS-MPI to get proper upper bound for BRTDP
* Add Dijkstra to get good lower bound for BRTDP
* Use BRTDP lower bound as heuristic for D* Lite

+++ osmnx_mdp/algorithms/cpp_brtdp_replan.cpp
@@ -16,7 +16,7 @@ void BRTDP_REPLAN::replan(const long &curr_node) {
 
 std::vector<long> BRTDP_REPLAN::get_path(
         google::dense_hash_map<long, long> &diverge_policy,
-        const float &beta,
+        const double &beta,
         const bool &always_replan) {
     // Defaults:
     //  beta = .02
